+++ packages/houdini-preprocess/package.json
@@ -18,7 +18,8 @@
 		"@types/jest": "^26.0.20",
 		"@types/mkdirp": "^1.0.1",
 		"babel-jest": "^26.6.3",
-		"houdini-compiler": "^0.0.15",
+		"houdini": "^0.0.15",
+		"houdini-common": "^0.0.15",
 		"jest": "^26.6.3",
 		"typescript": "^4.0.0"
 	},
