+++ packages/houdini/src/generators/runtime/template/index.ts
@@ -1,3 +1,5 @@
+import { GraphQLTagResult } from './types'
+
 export * from './environment'
 export * from './runtime'
 export * from './types'
@@ -8,7 +10,7 @@ export { default as fragment } from './fragment'
 
 // this template tag gets removed by the preprocessor so it should never be invoked.
 // this function needs to return the same value as what the preprocessor leaves behind for type consistency
-export function graphql(str: TemplateStringsArray): import('houdini-preprocess').GraphQLTagResult {
+export function graphql(str: TemplateStringsArray): GraphQLTagResult {
 	// if this is executed, the preprocessor is not enabled
 	throw new Error(
 		"Looks like you don't have the preprocessor enabled. Encountered it at runtime wrapping: \n " +
